[APPDEV-2354] getOrderCountForChefsTableEligibility is not using the orderhistory cache.
git-svn-id: https://appdevsvn.nj01/appdev/FreshDirect/trunk@17926 842219ff-9341-0410-a943-573460bdd4de

+++ projects/FDStore/src/com/freshdirect/fdstore/customer/FDCustomerManager.java
@@ -1414,7 +1414,7 @@ public class FDCustomerManager {
 	
 	public static int getOrderCountForChefsTableEligibility(FDIdentity identity) throws FDResourceException {
 		ErpOrderHistory history = getErpOrderHistoryInfo(identity);
-		return history.getOrderCountForChefsTableEligibility();
+		return (int)history.getOrderCountForChefsTableEligibility();
 	}
 	
 	public static double getOrderTotalForChefsTableEligibility(FDIdentity identity) throws FDResourceException {
